requirements.txt
django
djangorestframework
markdown
django-filter
django-cors-headers
django-rest-swagger
django-rest-framework-docs
django-rest-auth
requests
django-celery-results
django-celery
flower
